Moves new features out of gsim base - plust other fixes
