+++ doc/gsim/gsim_table.rst
@@ -1,0 +1,11 @@
+===========
+GMPE Tables
+===========
+
+.. module:: openquake.hazardlib.gsim.gsim_table
+
+.. autoclass:: AmplificationTable
+    :members:
+
+.. autoclass:: GMPETable
+    :members:
