+++ doc/gsim/index.rst
@@ -56,6 +56,7 @@ Built-in GSIMs
     garcia_2005
     geomatrix_1993
     ghofrani_atkinson_2014
+    gsim_table
     idriss_2014
     lin_lee_2008
     lin_2009
@@ -115,11 +116,3 @@ Calculation context
 
 .. autoclass:: DistancesContext
 
-GMPE Tables
------------
-
-.. autoclass:: AmplificationTable
-    :members:
-
-.. autoclass:: GMPETable
-    :members:
